+++ pkgs/dart_mcp_server/lib/src/mixins/analytics.dart
@@ -132,6 +132,11 @@ base mixin AnalyticsEvents
         if (errors.isNotEmpty) {
           result = CallToolResult(
             content: [
+              Content.text(
+                text:
+                    'Invalid tool arguments, make sure to read the schema '
+                    'and try again:',
+              ),
               for (final error in errors)
                 Content.text(text: error.toErrorString()),
             ],